requirements_deployment.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
numpy>=1.21.0
pillow>=9.4.0